+++ src/routes/email_sender.py
@@ -38,4 +38,4 @@ async def send_testemail(request: Request) -> Response:
     if succeeded:
         return Response(status_code=status.HTTP_204_NO_CONTENT)
 
-    return Response(status_code=status.HTTP_409_CONFLICT)
+    return Response(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)